Updata data
