+++ src/app/api/index.ts
@@ -1,2 +1,2 @@
 export * from './moviesApi';
-export * from './getDataFetch';
+export * from './data';
